+++ CHANGELOG.md
@@ -27,3 +27,5 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 - [Identity] Grant Permission REST API ([#6117](https://github.com/opensearch-project/OpenSearch/pull/6154))
 - [Identity] REST API for user GET and DELETE ([#6189](https://github.com/opensearch-project/OpenSearch/pull/6189))
 - [Identity] Extensions security setup and token creation ([#6204](https://github.com/opensearch-project/OpenSearch/pull/6204))
+- [Identity] REST API for user GET and DELETE ([#6189](https://github.com/opensearch-project/OpenSearch/pull/6189))
+- [Identity] REST API for Reset password ([#5990](https://github.com/opensearch-project/OpenSearch/pull/6309))